C++: distinguish template declarations and specializations when capturing template parameters
The original code doesn't distinguish them, so capturing template
parameters is failed if a specialization is in a declaration like:

   template <typename P> struct TestStruct<TemplateParameterStruct<P>> {};

When C++ parser finds '<typename', it allocates an array for capturing
template parameters, and it stores 'P' in ' P>' to the array.  However,
when the parser finds '<TemplateParameterStruct', the original code
clears the array, and prepares to capture template parameters in
'<TemplateParameterStruct<P>>'.  At this time 'P' is lost, and cannot
be tagged.

This change introduces multiple arrays for the purpose and manages
them with a stack. Identifiers found during parsing <...> are stored
in an array at the top of the stack. Looking up an identifier is done
over all arrays in the stack from the top.

So, for parsing '<typename P>', an array (X) is pushed to the stack.
'P' is stored in the array X. For parsing '<TemplateParameterStruct<P>>',
an array (Y) is pushed on array (X). The array (X) is not cleared.
'P' is not lost and can be tagged.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>