C++: distinguish template declaration and specialization when capturing template parameters
The original code doesn't distinguish them, so capturing template
parameters is failed if a specialization is in a declaration like:

   template <typename P> struct TestStruct<TemplateParameterStruct<P>> {};

For the input, C++ cannot capture P because when it finds
'<TemplateParameterStruct', an internal ptrArray tracking parameters
is reset; P is lost.

This change allows the array to be stackable. So an internal ptrArray
where P is stored is not cleared when the parser find
'<TemplateParameterStruct'; the parser uses a newly prepared ptrArray
during parsing '<TemplateParameterStruct<P>>'.

cxxTokenIsPresentInTemplateParameters() also is extended.  It scans
not only the current the internal ptrArray on the top of stack but
also scans all in the stack to know whether an identifier is a
template parameter or not.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>