main: use memory backed mio as a temporary file when ctags runs in sandbox mode
Conceptually this change reverts
eb47af965d5c03850cbdf01ef3b7ec99ed290bd4.

The change depends on a assumption; when "-o -" and "--sort=no"
options are given, ctags can run without a temporary file. However,
this assumption is never satisfied. A re-scanning parser, a parser
implementing `parser2' method of `parserDefinition', rewinrds the output
stream when the current parsing is failed. C/C++ parser implements
the `parser2' method. The output stream connects to a terminal
(specified with -o -) cannot rewind; ctags may get EPIEP error in such
case.

This change reintroduces a temporary file.  For avoiding calling extra
system calls, this change utilizes memory backed mio instead of
opening a new temporary file when ctags runs in the sandbox mode.  In
the sandbox mode, the kinds of system calls are limited.

NOTE: The impact of this change may not be small.  If we use corkQueue
effectively, we may be able to avoid rewinding output stream.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>