RpmSpec,Autoconf: parse %configure options
Autconf parser captures:

    ctags-libexecdir in "AC_ARG_WITH([ctags-libexecdir]," as optwith kind and
    tmpdir in "AC_ARG_ENABLE(tmpdir," as optenable kind.

These are captured as definitions tags.
This change is about reference tags for the kinds.

./configure script is invoked various places.
One of typical place is RPM spec file.
RPS spec files have a special notation (macro) for invoking
configure script: %configure ....

This change in RpmSpec parser scans the %configure lines.
If --enable-foo, --disable-bar, --with-baz and/or --without-bazz are
passed to %configure lines, the RpmSpec parser captures
foo, bar, baz, and/or bazz as reference tags of optwith or optenable
kinds.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>