main: fill extras field of pseudo tags
Consider a fragment of html file:

	<h1>!_TAG_PROGRAM_VERSION</h1>

our HTML parser captures it well with escaping as:

	\21_TAG_PROGRAM_VERSION ...

So readtags is not confused; readtags understands it is not a pseudo
tag. However, once the tag is stored to tagEntry data struct of
readtags, the client code cannot distinguish it with a pseudo tag
because the name of the tag is escaped when readtags stores it to the
data structure.

This change attaches "extras:pseudo" to pseudo tags.  This helps hits
whether a tag is a pseudo tag or not when a tag is stored to the data
structure; if "extras:pseudo" is not attached, the tag is not a pseudo
tag.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>