lregex: replace gnu_regex with Onigmo regex engine
Close #1861.

In the much of cases, I hope there is no impact on existing optlib
code using --regex-... option.

If `r` regex flag is given, you can use extended features of Onigmo
with ruby regex syntax.

A demonstration of one of miracle features:

;; input.mylang

    (define (f1) 1)
    (define ((f2)) #t)
    (define (((f3))) "abc")

    --langdef=mylang
    --map-mylang=.mylang
    --kinddef-mylang=f,fun,function, function returing a function, or function returing a function returing function...
    --_fielddef-mylang=symbol,symbol binding to the function
    --fields-mylang={symbol}
    --regex-mylang=/\(define +(([-a-z0-9]+)|\(\g<1>\))/\1/f/r{_field=symbol:\2}

See the r flag passed to --regex-mylang.

    (((f3)))	input.mylang	/^(define (((f3))) "abc")$/;"	f	symbol:f3
    ((f2))	input.mylang	/^(define ((f2)) #t)$/;"	f	symbol:f2
    (f1)	input.mylang	/^(define (f1) 1)$/;"	f	symbol:f1

Look at the name of tags. The pairs of `(' and `)' are balanced well.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>