Cxx: extend cxxParserExtractVariableDeclarations to return the corkIndex for the last committed tag
In the most cases, cxxParserExtractVariableDeclarations can attach "end:" fields to the
local variables tagged in the function.

However, if the terminator for a captured tag is not found in the
context of cxxParserExtractVariableDeclarations, the function cannot
attach the field to the tag. In such a case, its caller should attach the "end:"
field instead of the function.

To handle such a case, this change extends
cxxParserExtractVariableDeclarations to return the corkIndex for the
last committed tag to that "end:" field is not attached yet.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>