readtags: don't use printf alike functions
Currently tagsPrintTag() assumes a FILE of stdio as output stream.
This change is preparation for allowing to replace the FILE with a
user defined custom output stream.

This change replaces printf alike functions called in tagsPrintTag()
with fputs and putc. In the future, methods which the custom output stream
must define are only the replacement for the two: functions for printing a C
string and printing a character.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>