readtags: introduce --escape/-E option
As written in tags(5), characters like tab in a tag entry
should be escaped with backslash characters.

readtags can read such escaped tag entries.
It does "unescaping" when reading.
It prints unescaped data as is.

A client tool showing the output of readtags directly to a
user may want this behavior, printing unescaped data.

However, some of tools itself may not want
the behavior. This change is for such tools.
.e.g a tool merging two tags files into one.

This change intrdouces --escape/-E option. With enabling the option,
readtags prints the entries with escaping as written in tags(5).

Signed-off-by: Masatake YAMATO <yamato@redhat.com>