dsl,dsL/qualifier: don't pass tagEntry to DSLProc directly
In the future the qualifier (filter) engine and sorter engine may
share the implementation of operators like "eq?", "not", "substr?",
and so on.

The current implementation of operators assumes it takes only tagEntry
as an argument. This is o.k. for the qualifier engine.
However, it is not applicable to sorter engine.

This change extends DSLProc for taking arbitrarily arguments.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>