dsl: introduce true and false as aliases for #t and #f
Consider writing a client tool runs on emacs.
The tool may use -Q and -S options of readtags.

The developers of the client tool may want to pass
an S expression built-in emacs lisp to readtags
directly.

However, there was a gap in what we call S expression
in emacs and readtags.

readtags's S expression is scheme based; #t and #f is
used as representation of boolean values. These representation
is not used in emacs. Further more, `#' is special character
(prefix of a reader macro) in emacs. So, it is hard for emacs
to build an S expression passed to readtags if the expression
includes a boolean value.

To fill the gap, this change introduces `true' and `false'
values to the readtags side. So the client tool that runs on
emacs doesn't doesn't have to use #t and #f directly in S
expression building process.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>