Cxx: fix the memory leak related to field attaching
Close #2524.
The memory leak was introduced in fcc4853e.

When attaching a field to a tag, the original code calls
attachParserField() wrong way. The caller must own the field value
passed to attachParserField(); the caller must free the memory
object for the value after calling makeTagEntry().
However, the original code didn't do so. As the result, the memory
object for the value was leaked.

This change introduces the parser trash box. After calling
attachParserField(), the caller puts the memory object to the trash
box; the memory object is freed when parsing the current input file is
completed.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>