R: don't tag twice for the same function variables
R is a dynamic language; a variable is not declared explicitly.
If assigning a value to a variable and it is the first time, ctags
should capture it as a variable definition.

ctags doesn't and should not evaluate the code, so detecting the
first appearance of a variable is not easy. However, there are areas
we can improve.

* parameter

  If assigning a value to a variable listed in the function parameter
  list, it should not be tagged. Instead the parameter should be tagged.

* only the first assignment in the scope

  Only the first assignment to a variable in the scope should be tagged.
  The later assignments to the same variable in the same scope should be
  ignored.

This behavior is very different from the old R parser. The old R parser
tags all assignments.

Unlike JavaScript parser, the new R parser tags all assignments at the
top level if the values are functions.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>