readtags: don't overwrite const memory object
When making the value for $scope-kind, the const memory object
representing scope of an tag entry was overwritten.

It was needed to make an EsString object with es_string_new() that
takes a nul-terminated C string as initial value.

This change introduces es_string_newL() that takes a pointer to
a C string and its length for making an EsString object. It allows
making an EsString object from a substring of given C string.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>