[SELF-INCOMPATIBLE]readtags: don't apply csv separation when making values for $inherits and $roles
The values for $inherits and $roles were converted to a list by splitting
their raw values with commas. However, for some languages, just splitting
with commas is not enough.

Let's consider folllowing C++ input:

    template <typename T, typename S> class B {
      T t;
      S s;
    };

    class C
    {
      int y;
    };

    class A: B<int,float>, C
    {
      int x;
    };

tags file for the input:

     A	/tmp/foo.cpp	/^class A: B<int,float>, C$/;"	c	file:	inherits:B<int,float>,C
     ...

For the entry, the value for $inherits was ("B<int" "float>" "C").
This elements of list doesn't make sense.

Now the values for $inherits and $roles are provided as is.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>