dsl: handle :case-fold keyword for string->regexp correctly
("string->regexp" operator of readtags models "string->regexp"
procedure of gauche.)

readtags did case insensitive search when :case-fold is passed
as the second argument to "string->regexp" operator.

gauche does do the same when :case-fold and #t are passed
as the second and third arguments like

   (string->regex PATTERN :case-fold #f)

This change makes the argument handling of string->regexp
of readtags aligns to that of gauche.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>