writer-etags: add kind suffix to Ada's output
Close #2565.

etags handles Ada's output specially.
https://www.gnu.org/software/emacs/manual/html_node/emacs/Tag-Syntax.html

This change makes ctags do the similar as etags does.

NOTE: The Ada parser of ctags doesn't distinguish between a function
and a procedure, therefore a bit dirty technique is used in
writer-etags.  The value of the typeref field may be useful to
distinguish.  However, the field is not implemented in the parser.
Implementing the field in a TODO item.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>