main: allow memory allocators to return NULL when the size is 0
Close #2617.

man malloc(3):
    If size is 0, then malloc() returns either NULL, or a unique
    pointer value that can later be successfully passed to free().

The original code of eMalloc treats NULL returned from malloc as
"out of memory". This has been o.k. at least on GNU/Linux. On the
platform, malloc(0) return non-NULL value. However, on AIX, malloc(0)
return NULL. As the result ctags on AIX always stops with "out of
memory" in the very early stage of command execution.

With this change, ctags treats NULL returned from malloc as
"out of memory" only if the size is not 0.

This means NULL can be passed to eFree that doesn't accept
NULL. In such place, eFree invocations are replaced with
eFreeNoNullCheck that accept NULL.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>