CPreProcessor,refactor: introduce a struct representing parameters passed to cppInit()
cppInit takes too many arguments. To make the cppInit invocation
understandable, this change introduces cppInitData struct type. With
the type, a client parser can specify an argument with specifying the
name of field.

This change fixex a bug as a side effect.
The orders of parameters didn't matched between the declaration and
the implementation of cppInit().

Signed-off-by: Masatake YAMATO <yamato@redhat.com>