CPreProcessor,C: skip #ifdef __cplusplus ~ #endif branch
Close #2647.

C parser doesn't work when C++ code "extern "C" {" and "}" is given.

    $ cat /tmp/extern-c.c
    extern "C" {
    #ifdef X
      void f() {}
    #else
      void g() {}
    #endif
    }
    $ u-ctags --output-format=xref --kinds-c=+plz --fields=+nie -o - /tmp/extern-c.c
    f                function      3 /tmp/extern-c.c  void f() {}

    $ cat /tmp/none-extern-c.c
    #ifdef X
      void f() {}
    #else
      void g() {}
    #endif
    $ u-ctags --output-format=xref --kinds-c=+plz --fields=+nie -o - /tmp/none-extern-c.c
    f                function      2 /tmp/none-extern-c.c void f() {}
    g                function      4 /tmp/none-extern-c.c void g() {}

The function g() is missed if the code is surrounded by extern "C" { and }.

This change uses a heuristic rule to skip the code in the C parser.

In most of cases, the code is in #ifdef __cplusplus ~ #endif branch.
So when the CPreProcessor parser detects #ifdef __cplusplus ~ #endif,
with this change, the CPreProcessor parser doesn't pass the code inside __cplusplus branch
to its client parser that requests to the CPreProcessor parser to do so.

C parser requests it. C++ parser doesn't.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>