SystemVerilog: User defined type support
tagNameList: an identifier is classified as net or variable name when it
is followed by ',', ':', or ')'.

define processParameterList() for both module and class parameter list.

It is based on the code which was included in processClass().
tags are output in the order defined.  (class parameter was in reversed order.)
parens in expression was handled by skipPastMatch().
type with dimenssion is supported
better support of class inheritance

only identifiers followed by ',', ':', or ')' are output as nets or variables

labels are not passed to tagNameLlist() anymore.