main: make commonInParsers field of ptagDesc more generic
In many cases, ptags are categories to two groups.  One group is
parser specific. They hold parser specific values.  Typical one is
ptags about kinds.

Another group is common within parsers (parser neutral).

ctags recognized the category for a given ptag with calling
isPtagCommonInParsers() that return the commonInParsers field
of the ptag.

However, we found there is a case that calling the function is enough.
PTAG_FIELD_DESCRIPTION and PTAG_EXTRA_DESCRIPTION are in both groups.

To handle them, this change converts commonInParsers boolean field to
a bit flags; the first field is for representing whether the ptag
holds a parser specific value or not, and the second field is for
representing whether the ptag holds a value common within
parsers. PTAG_FIELD_DESCRIPTION and PTAG_EXTRA_DESCRIPTION set the
both bits.

isPtagCommonInParsers() can be used as before changing.
isPtagParserSpecific() is a new predicate.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>