main.c: fix invocation as “etags”
Call initializeParsing() before testEtagsInvocation() so that LanguageTable
is set up.

This fixes a crash because of NULL pointer dereference: if LanguageTable is
not initialized, then testEtagsInvocation() calls setEtagsMode(), which when
the binary is called "etags", calls enableLanguage(), which crashes.

In other words, this allows "ctags" to be called "etags" and run
successfully.