SystemVerilog: skip paired parens and braces in the main loop in findVerilogTags()
This fix makes the token processing much more reliable and robust.
It also ignores locally declared variables in a for-loop.
It looks like a drawback, but from a user's point of view
they don't need to be tagged and they are rather noise.

Units(SystemVerilog): ignore locally declared variables in a for-loop