Move EXTERNAL_PARSER_LIST ifdef to the beginnning of BuiltInParsers
Commit

82791c20ff8d01e7055bcbbb30db4999d2c3d463

moved CTagsParser, FallbackParser and CTagsSelfTestParser above the
ifdef which made these parsers included also when using
EXTERNAL_PARSER_LIST - this is not good for Geany because these three
parsers are "invisible" for it and shift the EXTERNAL_PARSER_LIST by 3.

This patch moves the ifdef back where it was before.