main,units.py: introduce "epoch:" field for F/file kind
For implementing the feature "updating the tags file" in the future,
this change introduces "epoch:", a field which records the last
modified time of the input file. This field is enabled by default.

Example output:

    $ ./ctags --extras=f --kinds-C= -o - main/main.c
    main.c	main/main.c	1;"	F	language:C	epoch:1078071702

The resolution of time is second in the current implementation.
When using higher resolution timestamp in the future, we will introduce "epochLow:" or
something new field for representing the lower part of epoch. So the epoch field
holds an integer even in the future. However, we cannot promise tbe bit length of
the integer.

The epoch fields in tags output makes many test cases in Units not work.
So this change puts "--fileds=-{epoch}" to units.py for turning off
the field when running "make units".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>