R: use "function" kind for functions defined in a function
<BUG>
The original code attaches "funcvar" kind to functions defined in a
function.

If a variable in a function gets a value returned from a well-known
constructor, R parser should attach a suitable kind for the
constructor to the variable.

In this case, function() is the well-known constructor.

<HOW IT SHOULD BE>
"function" kind is for a variable that has a value returned from
function(), or a value itself returned from function(). In the latter
case, ctags generates an anonymous name for tagging the value.

"function" kind should be used whether the variable is defined
in the top-level context or in a function.

<NOTE>

Let's call function() a well-known constructor.

If a variable in a function has a value returned from a function call
OTHER THAN the well-known constructor, the tag for the variable should
have "globalVar" or "functionVar" as its kind.  If the variable is
assigned in the top-level context, it should have "globalVar" kind. If
not, it should have "functionVar" kind.

In the future, R parser may support c(), list(), and data.frame() as
well-known constructors. We will introduce "vector",
"list" and "daraFrame" kinds for them

This change is not only a bug fix but also a preparation for #2825.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>