Tmain: test warning messages only when an input file is given
Whether disabling a field is acceptable or not is upto choice of a
writer. The ctags writer has some fixed fields.

The choice is decided when an input file is given.  The original code
specifies --_force-quit but it doesn't specify an input file.
This test case has worked because of a bug in the main part.
To avoid the trouble arised when fixing the bug, this change
specifies an input file explicitly.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>