main: relax the order of --output-format= and --fields= options in a cmdline
Close #2860.

Quoted from #2860.

    $ ctags --output-format=json --fields=+n-P -o - qualified.java
    {"_type": "tag", "name": "point", "path": "qualified.java", "line": 1, "kind": "class"}
    {"_type": "tag", "name": "x", "path": "qualified.java", "line": 2, "kind": "field", "scope": "point", "scopeKind": "class"}
    $ ctags --fields=+n-P --output-format=json -o - qualified.java
    ctags: Warning: Cannot disable fixed field: 'P'{pattern}
    {"_type": "tag", "name": "point", "path": "qualified.java", "pattern": "/^class point {$/", "line": 1, "kind": "class"}
    {"_type": "tag", "name": "x", "path": "qualified.java", "pattern": "/^\tdouble x;$/", "line": 2, "kind": "field", "scope": "point", "scopeKind": "class"}
    $

Whether disabling a field is acceptable or not is up to choice of a
writer. The ctags writer has some fixed fields but the json writer
doesn't have any such fields.

However, the default writer is the ctags writer. So disabling
a fixed field before specifying a writer causes a warning.
The original code warned when disabling a fixed field.
This change delays making warnings till starting parsing.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>