main: adjust states of enabled/disabled for fixed fields before listing with --list-fields
This change is a continuation of c0421c5ec596b5f2a3e1a16a61d6ee3368a68842.

The effect of the "FIXED" property of fields cannot be decided till ctags
chooses of an output writer.

c0421c5ec596b5f2a3e1a16a61d6ee3368a68842 considers beginning
of parsing is the timing of the choice.

c0421c5ec596b5f2a3e1a16a61d6ee3368a68842 misses there is one more
case: the choice is made when --list-fields option is processed.

To reflect the choice to the output of --list-fields, this
change calls writerCheckOptions at the beginning of the process.

With this change, the behavior of --list-fields option
consists with the following description of ctags(1):

    ``--list-fields[=language|all]``
	    ...
	    FIXED
		    Whether this field can be disabled or not in tags output.

		    Some fields are printed always in tags output.
		    They have ``yes`` as the value for this column.

		    Unlike the tag output mode, JSON output mode allows disabling
		    any fields.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>