Ada: skip string literal in a case
Close #2913.

Ada parser has some functions for skipping input stream.
They don't consider string literals in the stream.

This change fixes the bug in a very limited area.

    ...
	package Generic1 is new Generic_G (Description => +"; " +":");
    ...

adaParseBlock() tries skipping to ';' from "is".  In this case, as the
result, adaParseBlock() expects the input stream points ';' at the end
of line. However, the expectation was not satisfied because there was a
string literal that included ';' between "new" and ';' at the end of
line.

This change extends skipPastLambda() to skip string literal.
adaParseBlock() calls skipPastLambda() indirectly via
skipPastKeywordOrWord(). As the result, the expectation is satisfied.

Extending other functions for skipping may be done when we get more
bug reports.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>