main: don't put a newly allocated struct member to parser trash box if the tag in corkQueue
The original code puts the newly allocated buffer for extraDynamic
member of tagEntryInfo struct to the parser trash box.

The original code works fine when calling markTagExtraBit()
for a tagEntryInfo on the stack.

However, the original code causes a carsh when calling the functions
for a tagEntryInfo in a corkQueue. When freeing a tagEntryInfo in a
corkQueue, the destructor frees the member. Here is a double free
causing the crash: via the destructor and trashBoxDelete().

This change introduces a bit flag. inCorkQueue to tagEntryInfo structure.
The bit is set when a tagEntryInfo is put to the corkQueue.

markTagExtraBit() puts the member only if the bit is not set.

This change will be tested when extras: operator.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>