[INCOMPATIBLE]main: don't read CTAGS and/or ETAGS environment variables
Close #2890.

Using these variables as source of configuration could be trouble.

Consider the following Makefile:

  CTAGS=/usr/bin/ctags

  tags: $(SOURCE_FILES)
  	$(CTAGS) $?

People can use the Makefile with the following command line:

   $ export CTAGS=/home/yamato/bin/u-ctags
   $ make CTAGS=$CTAGS

The user may get a strange warning message:

   Ignoring non-option in CTAGS variable

It is hard for users to understand the meaning of the warning message
if the user doesn't know the variable is used ctags itself as the
source of configuration.

Even u-ctags developers take long time to understand what happened
when #2048 is reported.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>