SQL: run guest parsers on strings where procedures are defined with procedural languages
Close #3017.

PostgreSQL allows users defining functions and procedures with languages other than SQL.
Languages for defining a procedures or functions are called procedural languages.

From the view point of the syntax of the base language (SQL), the definitions are
just strings.

This change recognizes the strings and LANGUAGE keywords, and runs
guest parsers on the strings if syntactical conditions are met.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>