[SELF-INCOMPATIBLE]C++: introduce "used" role to "name" kind
WARNING: This change extracts "bar" in "using foo::bar" as a referred tag,
         not a definition tag.

"bar" in "using foo::bar" is captured as a definition tag of "name" kind.
However, "foo" is not defined in the statement.

"foo" is referred as a something defined "bar".

Therefore, ctags should not capture "bar" as a definition tag.
Instead, ctags should capture it as a reference tag.

The original code captures "bar" as a definition tag of "name" kind.
This change captures "bar" as a reference tag of "used" role of
"name" kind.

  $ cat /tmp/foo.hh
  using foo::bar;
  $ ./ctags -o - /tmp/foo.hh
  $ ./ctags -o - --extras=+r /tmp/foo.hh
  bar	/tmp/foo.hh	/^using foo::bar;$/;"	N
  $ ./ctags -o - --extras=+r --fields=+K /tmp/foo.hh
  bar	/tmp/foo.hh	/^using foo::bar;$/;"	name
  $ ./ctags -o - --extras=+r --fields=+Kr /tmp/foo.hh
  bar	/tmp/foo.hh	/^using foo::bar;$/;"	name	roles:used

  $ ./ctags --list-roles=C++.'{name}'
  #KIND(L/N) NAME ENABLED DESCRIPTION
  N/name     used on      specified with "using foo::bar"

Signed-off-by: Masatake YAMATO <yamato@redhat.com>