lua: don't include module in function names
The name of a module in a Lua source file is, in most cases, not what is
uses by other Lua modules. For example,

test.lua:

    local M = {}

    function M.foo()

    return M

my_module.lua:

    local test = require("test")

    test.foo()

ctags includes the module namespace in the tag output ('M.foo' rather
than simply 'foo'), which makes the tag much less useful when used in
other Lua files.

Instead, make the tag output only include the identifier of the function
itself without any of the module namespace hierarchy. To accomplish
this, we reset the identifier string buffer upon encountering either a
'.' or ':' character. We also update 'isLuaIdentifier' to exclude these
two characters (this is necessary for 'extract_prev_token' which moves
backward character-wise until 'isLuaIdentifier' returns false).