Lua: fill scope fields in tags of function kinds
With the last commit for the Lua parser, "lua: don't include module in function names",
the lua parser degraded in the following aspects.

    test = {}
    test.f = function()
	   pirint "hello"
    end

The original parser can represent the relation between "test" and "f";
the "f" is extracted as "test.f".

With the last commit, the parser cannot represent the relation between
"test" and "f".

To recover the degradation, this change fills the scope fields as following:

   f	iput.lua	/^test.f = function()$/;"	kind:function	scope:unknown:test
   test.f	input.lua	/^test.f = function()$/;"	kind:function	scope:unknown:test	extras:qualified

"test.f" is tagged only when --extras=+q is given.

Far from perfect but better than nothing.
TODO: tagging "test".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>