TclOO: introduce "forceUse" parameter
The original code expects the pattern `namespace import oo' in input.
The pattern is used for detecting whether TclOO is used in the input file.
However, there can be a case that the pattern is not appeared though
the input file uses TclOO:

   namespace import oo::*
   source input.tcl

To allow users to handle the none-self-contained case manually, this
change introduces the parameter `--param-TclOO.forceUse=1`.  If the
parameter is set, the parser runs as if 'namespace import oo::*' is
specified at the head of input files.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>