ITcl: introduce "forceUse" parameter
The original code expects the pattern `namespace import itcl' in input.
The pattern is used for detecting whether Itcl is used in the input file.
However, there can be a case that the pattern is not appeared though
the input file uses Itcl:

   package require Itcl
   namespace import itcl::*
   source input.tcl

To allow users to handle the none-self-contained case manually, this
change introduces the parameter `--param-Itcl.forceUse=1`.  If the
parameter is set, the parser runs as if 'namespace import itcl::*' is
specified at the head of input files.

Suggested-by: @bigfaceworm
Signed-off-by: Masatake YAMATO <yamato@redhat.com>