VHDL: avoid dereferencing NULL as a tagEntryInfo
Close #3158.

In the process of finding the end of a package, the parser accessed
the internal of the tag entry for the package for filling end: filed
of the tag.  In the case the tag of the package was not created, the
accessing caused a crash. This change limits the access.

This change doesn't consider the reason why the tag for the package is
not created.

Following code snippets taken from the real code are triggering
the crashes:

    -- psi_fix/model/snippets/psi_fix_pkg_writer_tmpl.vhd
    -- in https://github.com/paulscherrerinstitute/psi_fix
    -- reported in #3158 by @tsenart
    package body <PACKAGE_NAME> is
    ...
    end <PACKAGE_NAME>;

    -- ghdl/testsuite/vests/vhdl-ams/ashenden/compliant/access-types/ordered_collection_adt.vhd
    -- in https://github.com/peteut/ghdl
    -- reported in #3158 by @tsenart
    package «element_type_simple_name»_ordered_collection_adt is
    ...

Signed-off-by: Masatake YAMATO <yamato@redhat.com>