optlib2c: introduce --__selector-<LANG>= pseudo option for injecting a select to optlib parser
LEX parser, a parser in development, has .l in its map.  It conflicts
with Lisp parser that has also .l in its map.

A selector is needed for arbitration.

As a temporarily solution, this change introduces --__selector-<LANG>=
option to optlib2.  optlib2c sets the C language function name
specified with the option to the selectLanguage member of the Parser
structure.

In addition, this change introduces a new convention; ctags ignores
an argument started with '--__'. optlib2c may handle the argument.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>