main: consider BOM in offset and line calculation for making a promise
A multi-table regex parser parses a MIO where all the content of a
file are loaded.

A MIO may include BOM. However, before starting parsing the file head
is moved over the BOM. So the parser never sees the BOM.

The infrastructure code for implementing multi-table regex parsers
tracks and manages the file offset in the layer(lregex). The offset is used
when making a promise to run a guest parser.

There is a gap between the offset in the MIO level and the lregex
level. In the MIO level, the offset includes the BOM. On the other hand,
in the lregex level, the offset doesn't include the BOM; the offset is
relative to the end of BOM.

===========V: the MIO level
[B][O][M]ABC
         ==^: the lregex level

Because the gap, a multiple table parser may make a promise
having wrong offsets if the input file started from a BOM.

This change fills the gap by considering the BOM in the
functions for converting between offsets and lines.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>