main: make automatic FQ tag emission based on the language associated with the target tag
A language can specify whether the main part does automagic-FQ-tag emission or not
with requestAutomaticFQTag field of parserDefinition.

To access the requestAutomaticFQTag field, the original code calls
getInputLanguage(). However, this didn't work expectedly. When a
subparser made a tag, it was stored to the corkQueue. The corkQueue
was flushed after the subparser run, within the base parser run.
As the result requestAutomaticFQTag field of the base parser was
referred when calling getInputLanguage().

This changes refers the tag itself to access the requestAutomaticFQTag
field instead of calling getInputLanguage().

Signed-off-by: Masatake YAMATO <yamato@redhat.com>