Markdown: Improve handling of underlined headings
Headings defined using underline by either '=' or '-' may contain an
arbitrary amount of '=' or '-' characters which don't have to be equal
to the number of characters on the previous line (which was what the
previous code did).

Special attention has to be paid to code blocks using indentation
(either by 4 or more spaces or a \t character). For instance

    abcd
---

this is a code block followed by underline, while

  abcd
---

is a 'abcd' heading.