main: adjust the end fields of tags that scopes are cleared with "set" action
'/regex/\n/x/{scope=set}' clears the scope stack, then pushes the newly
created tag on the stack.

The original code set the start line of the newly created tag to the end
fields of the tags pushed on the scope stack when clearing.

The new code sets (the start line of the newly created tag) minus one
to the field.

Let's think about the following input (input.x) for an imaginary language X:

    1: namespace NS1:
    2:   def x = 1
    3:
    4: namespace NS2:
    5:   def y = 1
    6:

and think about the optlib file for the X:

    --langdef=X
    --map-X=.x
    --kinddef-X=n,namespace,namespaces
    --kinddef-X=v,variable,variables
    --regex-X=/namespace +([A-Z][A-Z0-9]*)/\1/n/{scope=set}
    --regex-X=/ *def +([a-z])/\1/v/{scope=ref}

The original code emitted:

    NS1	input.x	/^namespace NS1:$/;"	n	end:4
    x	input.x	/^  def x = 1$/;"	v	namespace:NS1
    NS2	input.x	/^namespace NS2:$/;"	n	end:6
    y	input.x	/^  def y = 1$/;"	v	namespace:NS2

The code with this change emits:

    NS1	input.x	/^namespace NS1:$/;"	n	end:3
    x	input.x	/^  def x = 1$/;"	v	namespace:NS1
    NS2	input.x	/^namespace NS2:$/;"	n	end:6
    y	input.x	/^  def y = 1$/;"	v	namespace:NS2

The end field for `NS1` is adjusted as expected.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>