ada: remove setjmp() longjmp()
setjmp() and longjmp() may cause memory leaks and are better to avoid.
The ada parser actually doesn't seem to use them normally, there are proper
checks for EOF in the code, and longjmp() is used only when something
goes wrong after 1000 hits of EOF.

I went through all the loops and all of them seem to check for EOF so
there should be no need for setjmp()/longjmp().