main: experiment to detect memory access behind EOL
Currently when reading the input file, ctags reads it line by line and
maintains a (always growing) vString value for the current line where
characters of the line are stored. Since the string never shrinks,
it may be very long and may contain valid code from previous lines
behind the '\0' character. If parsers contain a bug and read behind
the terminating '\0', this may be undetected by valgrind because
the string is legally allocated.

This experiment re-allocates the string after reading a line to the
exact size of the line which should make valgrind complain if the string
is accessed behind the terminating '\0' by some parser.

Not sure why but the allocated buffer has to be at least 2 bytes long,
otherwise tests start to fail.