Yaml,OpenAPI: make the OpenAPI parser's stack related code reusable
The technique used in the OpenAPI may be helpful to implement another YAML subparser.
So, I moved the code implementing the technique to the YAML parser experimentally
in this change.

The original code uses an array of keywords for pointing out interesting tokens.
This change uses a string for the same purpose. The routines in this change
convert the string to an array of keywords.
Till a real specification and its implementation becomes available, we call
the format of the string Ypath.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>